+++ Drivers/Nano-OpenTherm/Requests/GetBoilerTemperatureRequest.cs
@@ -7,8 +7,8 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Requests
         public GetBoilerTemperatureRequest() : base() { }
         public GetBoilerTemperatureRequest(Request baseReq) : base(baseReq) { }
 
-        protected override ulong GetRawDataCore() => ProcessRequest(0);
-        protected override void SetRawDataCore(ulong value) { /* no properties to populate */ }
+        protected override uint GetRawDataCore() => ProcessRequest(0);
+        protected override void SetRawDataCore(uint value) { /* no properties to populate */ }
         public override MessageType MessageType => MessageType.READ_DATA;
         public override MessageID MessageID => MessageID.Tboiler;
     }
