+++ Drivers/Nano-OpenTherm/Requests/GetCumulativeElectricityProductionRequest.cs
@@ -9,8 +9,8 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Requests
 
         public ushort KWh { get; set; }
 
-        protected override ulong GetRawDataCore() => ProcessRequest(KWh);
-        protected override void SetRawDataCore(ulong value) { KWh = Utilities.GetLowUShort(value); }
+        protected override uint GetRawDataCore() => ProcessRequest(KWh);
+        protected override void SetRawDataCore(uint value) { KWh = Utilities.GetLowUShort(value); }
 
         public override MessageType MessageType => MessageType.READ_DATA;
         public override MessageID MessageID => MessageID.CumulativElectricityProduction;
