+++ Drivers/Nano-OpenTherm/Requests/GetPressureRequest.cs
@@ -7,8 +7,8 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Requests
         public GetPressureRequest() : base() { }
         public GetPressureRequest(Request baseReq) : base(baseReq) { }
 
-        protected override ulong GetRawDataCore() => ProcessRequest(0);
-        protected override void SetRawDataCore(ulong value) { }
+        protected override uint GetRawDataCore() => ProcessRequest(0);
+        protected override void SetRawDataCore(uint value) { }
         public override MessageType MessageType => MessageType.READ_DATA;
         public override MessageID MessageID => MessageID.CHPressure;
     }
